move render mcp prompt into its own file
